Renamed @ModuleChannels to @Binding, moved Source, Sink, Processor
