+++ counter-sink/src/main/java/org/springframework/cloud/stream/module/metrics/CounterSink.java
@@ -21,7 +21,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.metrics.CounterService;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Sink;
+import org.springframework.cloud.stream.messaging.Sink;
 import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.messaging.Message;
 
@@ -30,6 +30,7 @@ import org.springframework.messaging.Message;
  *
  * @author Eric Bottard
  * @author Mark Pollack
+ * @author Marius Bogoevici
  */
 @EnableBinding(Sink.class)
 public class CounterSink {
