+++ filter-processor/src/main/java/org/springframework/cloud/stream/module/filter/FilterProcessor.java
@@ -19,8 +19,8 @@ package org.springframework.cloud.stream.module.filter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Processor;
 import org.springframework.cloud.stream.config.SpelExpressionConverterConfiguration;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.context.annotation.Import;
 import org.springframework.integration.annotation.Filter;
 import org.springframework.messaging.Message;
@@ -29,6 +29,7 @@ import org.springframework.messaging.Message;
  * A Processor module that retains or discards messages according to a predicate.
  *
  * @author Eric Bottard
+ * @author Marius Bogoevici
  */
 @EnableBinding(Processor.class)
 @Import(SpelExpressionConverterConfiguration.class)
