+++ ftp-source/src/main/java/org/springframework/cloud/stream/module/ftp/FtpSource.java
@@ -19,9 +19,9 @@ import java.util.Collections;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.cloud.stream.annotation.Bindings;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
-import org.springframework.cloud.stream.annotation.Source;
+import org.springframework.cloud.stream.messaging.Source;
 import org.springframework.cloud.stream.module.MaxMessagesProperties;
 import org.springframework.cloud.stream.module.PeriodicTriggerConfiguration;
 import org.springframework.cloud.stream.module.file.FileConsumerProperties;
@@ -43,6 +43,7 @@ import org.springframework.scheduling.Trigger;
 
 /**
  * @author David Turanski
+ * @author Marius Bogoevici
  */
 @EnableBinding(Source.class)
 @EnableConfigurationProperties({FtpSourceProperties.class,
@@ -66,7 +67,7 @@ public class FtpSource {
 	Trigger trigger;
 
 	@Autowired
-	@ModuleChannels(FtpSource.class)
+	@Bindings(FtpSource.class)
 	Source source;
 
 	@Bean
