+++ groovy-filter-processor/src/test/java/org/springframework/cloud/stream/module/filter/GroovyFilterProcessorApplicationIntegrationTests.java
@@ -27,8 +27,8 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
-import org.springframework.cloud.stream.annotation.Processor;
+import org.springframework.cloud.stream.annotation.Bindings;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.MessageCollector;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.GenericMessage;
@@ -39,6 +39,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
  * Integration Tests for GroovyFilterProcessor.
  *
  * @author Eric Bottard
+ * @author Marius Bogoevici
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = GroovyFilterProcessorApplication.class)
@@ -47,7 +48,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 public abstract class GroovyFilterProcessorApplicationIntegrationTests {
 
 	@Autowired
-	@ModuleChannels(GroovyFilterProcessor.class)
+	@Bindings(GroovyFilterProcessor.class)
 	protected Processor channels;
 
 	@Autowired
