+++ groovy-transform-processor/src/main/java/org/springframework/cloud/stream/module/transform/GroovyTransformProcessor.java
@@ -19,7 +19,7 @@ package org.springframework.cloud.stream.module.transform;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Processor;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.module.common.ScriptVariableGeneratorConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
@@ -34,6 +34,7 @@ import org.springframework.scripting.support.ResourceScriptSource;
  *
  * @author Eric Bottard
  * @author Mark Fisher
+ * @author Marius Bogoevici
  */
 @EnableBinding(Processor.class)
 @Import(ScriptVariableGeneratorConfiguration.class)
