+++ groovy-transform-processor/src/test/java/org/springframework/cloud/stream/module/transform/GroovyTransformProcessorIntegrationTests.java
@@ -26,8 +26,8 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
-import org.springframework.cloud.stream.annotation.Processor;
+import org.springframework.cloud.stream.annotation.Bindings;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.MessageCollector;
 import org.springframework.messaging.support.GenericMessage;
 import org.springframework.test.annotation.DirtiesContext;
@@ -37,6 +37,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
  * Integration Tests for the Groovy Transform Processor.
  *
  * @author Eric Bottard
+ * @author Marius Bogoevici
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = GroovyTransformProcessorApplication.class)
@@ -45,7 +46,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 public abstract class GroovyTransformProcessorIntegrationTests {
 
 	@Autowired
-	@ModuleChannels(GroovyTransformProcessor.class)
+	@Bindings(GroovyTransformProcessor.class)
 	protected Processor channels;
 
 	@Autowired
