+++ log-sink/src/main/java/org/springframework/cloud/stream/module/log/LogSink.java
@@ -20,11 +20,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Sink;
+import org.springframework.cloud.stream.messaging.Sink;
 import org.springframework.integration.annotation.ServiceActivator;
 
 /**
  * @author Dave Syer
+ * @author Marius Bogoevici
  */
 @EnableBinding(Sink.class)
 public class LogSink {
