+++ redis-sink/src/main/java/org/springframework/cloud/stream/module/redis/sink/RedisSink.java
@@ -19,7 +19,7 @@ package org.springframework.cloud.stream.module.redis.sink;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Sink;
+import org.springframework.cloud.stream.messaging.Sink;
 import org.springframework.integration.annotation.ServiceActivator;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageHandler;
@@ -29,6 +29,7 @@ import org.springframework.messaging.MessageHandler;
  *
  * @author Eric Bottard
  * @author Mark Pollack
+ * @autor Marius Bogoevici
  */
 @EnableBinding(Sink.class)
 public class RedisSink {
