+++ time-source/src/main/java/org/springframework/cloud/stream/module/time/TimeSource.java
@@ -22,7 +22,7 @@ import java.util.Date;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Source;
+import org.springframework.cloud.stream.messaging.Source;
 import org.springframework.cloud.stream.module.PeriodicTriggerConfiguration;
 import org.springframework.context.annotation.Import;
 import org.springframework.integration.annotation.InboundChannelAdapter;
@@ -31,6 +31,7 @@ import org.springframework.integration.annotation.Poller;
 /**
  * @author Dave Syer
  * @author Glenn Renfro
+ * @author Marius Bogoevici
  */
 @EnableBinding(Source.class)
 @EnableConfigurationProperties(TimeSourceProperties.class)
