+++ time-source/src/test/java/org/springframework/cloud/stream/module/time/TimeSourceIntegrationTests.java
@@ -24,8 +24,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.IntegrationTest;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
-import org.springframework.cloud.stream.annotation.Source;
+import org.springframework.cloud.stream.annotation.Bindings;
+import org.springframework.cloud.stream.messaging.Source;
 import org.springframework.cloud.stream.test.binder.MessageCollector;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -40,7 +40,7 @@ import static org.junit.Assert.assertEquals;
 public abstract class TimeSourceIntegrationTests {
 
 	@Autowired
-	@ModuleChannels(TimeSource.class)
+	@Bindings(TimeSource.class)
 	protected Source timeSource;
 
 	@Autowired
