+++ transform-processor/src/main/java/org/springframework/cloud/stream/module/transform/TransformProcessor.java
@@ -19,8 +19,8 @@ package org.springframework.cloud.stream.module.transform;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.cloud.stream.annotation.EnableBinding;
-import org.springframework.cloud.stream.annotation.Processor;
 import org.springframework.cloud.stream.config.SpelExpressionConverterConfiguration;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.context.annotation.Import;
 import org.springframework.integration.annotation.Transformer;
 import org.springframework.messaging.Message;
@@ -29,6 +29,7 @@ import org.springframework.messaging.Message;
  * A Processor module that transforms messages using a SpEL expression.
  *
  * @author Eric Bottard
+ * @author Marius Bogoevici
  */
 @EnableBinding(Processor.class)
 @Import(SpelExpressionConverterConfiguration.class)
