+++ transform-processor/src/test/java/org/springframework/cloud/stream/module/transform/TransformProcessorIntegrationTests.java
@@ -26,8 +26,8 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.cloud.stream.annotation.ModuleChannels;
-import org.springframework.cloud.stream.annotation.Processor;
+import org.springframework.cloud.stream.annotation.Bindings;
+import org.springframework.cloud.stream.messaging.Processor;
 import org.springframework.cloud.stream.test.binder.MessageCollector;
 import org.springframework.messaging.support.GenericMessage;
 import org.springframework.test.annotation.DirtiesContext;
@@ -37,6 +37,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
  * Integration Tests for the Transform Processor.
  *
  * @author Eric Bottard
+ * @author Marius Bogoevici
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringApplicationConfiguration(classes = TransformProcessorApplication.class)
@@ -45,7 +46,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 public abstract class TransformProcessorIntegrationTests {
 
 	@Autowired
-	@ModuleChannels(TransformProcessor.class)
+	@Bindings(TransformProcessor.class)
 	protected Processor channels;
 
 	@Autowired
